docs: add documentation for client router
