+++ packages/hydrogen/src/components/Link/index.ts
@@ -1,1 +1,1 @@
-export {Link, useNavigate} from './Link.client';
+export {Link} from './Link.client';
