+++ packages/hydrogen/src/hooks/useNavigate/index.ts
@@ -1,0 +1,1 @@
+export {useNavigate} from './useNavigate';
