+++ packages/hydrogen/src/hooks/useNavigate/useNavigate.ts
@@ -1,0 +1,21 @@
+import {useRouter} from '../../foundation/Router/Router.client';
+
+type NavigationOptions = {
+  replace?: boolean;
+  reloadDocument?: boolean;
+  clientState?: any;
+};
+
+export function useNavigate() {
+  const router = useRouter();
+
+  return (
+    path: string,
+    options: NavigationOptions = {replace: false, reloadDocument: false}
+  ) => {
+    // @todo wait for RSC and then change focus for a11y?
+    if (options?.replace)
+      router.history.replace(path, options?.clientState || {});
+    else router.history.push(path, options?.clientState || {});
+  };
+}
